IISIM 2.6.1: Performance/Stability Debugging Removed redundant type casting (float/double) - Some casting remained from before all drawing used double - e.g. PointD used to be PointF Implemented scaling draw resolution for ECG_Isoelectric() and Add_Beat__Cardiac_Baseline() - Hyper-resolution of flat lines causing heap overflow #208 - Now HR interval of .75 - 5 seconds -> resolution coefficient of 1 - 100   - Resolution of 10ms - 1000ms; is also smoothed by forwardBuffer!

+++ II Library/Classes/Waveform.Draw.cs
@@ -76,9 +76,11 @@ namespace II.Waveform {
 
         /*
          * Generic flat-line
+         * Note: Resolution is a coefficient for draw time resolution in milliseconds; e.g. base
+         * ResolutionTime (default 10) * _Resolution (2) = 20ms between each PointD plotted
          */
 
-        public static List<PointD> Flat_Line (double _Length, double _Isoelectric, int _Resolution = 1) {
+        public static List<PointD> Flat_Line (double _Length, double _Isoelectric, double _Resolution = 1) {
             return Plotting.Line ((int)(ResolutionTime * _Resolution), _Length, _Isoelectric, new PointD (0, _Isoelectric));
         }
